pyproject.toml
[project]
name = "bibipay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "django>=5.2.7",
    "django-widget-tweaks>=1.5.0",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
]
